requirements.txt
asgiref==3.8.1
Django==5.1
gunicorn==23.0.0
packaging==24.1
pillow==10.4.0
sqlparse==0.5.1
tzdata==2024.1
whitenoise==6.7.0
